add middleware for protecting routes

+++ routes/web.php
@@ -17,24 +17,35 @@ use Illuminate\Support\Facades\Route;
 
 Route::get("/", [ListingController::class, "index"]);
 
-Route::get("/listings/create", [ListingController::class, "create"]);
+// Show the create form for creating a gib
+Route::get("/listings/create", [ListingController::class, "create"])->middleware("auth");
 
+// Show a single listing or gig
 Route::get("/listings/{listing}", [ListingController::class, "show"]);
 
-Route::post("/listings", [ListingController::class, "store"]);
+// Create a listing
+Route::post("/listings", [ListingController::class, "store"])->middleware("auth");
 
-Route::get("/listings/{listing}/edit", [ListingController::class, "edit"]);
+// Shows the form for editing a listing or gig
+Route::get("/listings/{listing}/edit", [ListingController::class, "edit"])->middleware("auth");
 
-Route::put("/listings/{listing}", [ListingController::class, "update"]);
+// Update a listing or gig
+Route::put("/listings/{listing}", [ListingController::class, "update"])->middleware("auth");
 
-Route::delete("/listings/{listing}", [ListingController::class, "destroy"]);
+// Delete a listing or gig
+Route::delete("/listings/{listing}", [ListingController::class, "destroy"])->middleware("auth");
 
-Route::get("/register", [UserController::class, "create"]);
+// Shows the form for registering as a new user
+Route::get("/register", [UserController::class, "create"])->middleware("guest");
 
+// Create new user
 Route::post("/users", [UserController::class, "store"]);
 
+// Logout user
 Route::post("/logout", [UserController::class, "logout"]);
 
-Route::get("/login", [UserController::class, "login"]);
+// Login user
+Route::get("/login", [UserController::class, "login"])->name("login")->middleware("guest");
 
+// Authenticate user
 Route::post("/users/authenticate", [UserController::class, "authenticate"]);
